Particles modification Particles time is measured in seconds, added spark particle when debrils colides with shield.

+++ Traction/src/dev/aaronps/traction/ParticleSystem.java
@@ -4,6 +4,6 @@ import android.graphics.Canvas;
 
 public interface ParticleSystem
 {
-	public void logic(final long time);
+	public void logic(final float time);
 	public void draw(final Canvas canvas);
 }
